Speeds up counter reads in pps-gmtimer.
The normal omap_dm_timer_read_counter() function checks for "posted"
mode (which is normally set by default), and if it's set, checks the
posted-write status register for a posted write to the counter.  But
since we never actually write the counter, this is just a useless
extra timer register read adding delay.

This change creates a local counter-read function which uses the
lower-level __omap_dm_timer_read_counter() function with the "posted"
argument supplied as 0.  It's also inline, saving a call to an
external function, and the optimizer will presumably remove the
unreachable posting-check code.

It also changes the capture-register read to disable the posting
check, which never did anything, anyway.

TESTED:
Verified that the PPS devices still work.
Observed about a 1/3 reduction in ts_spread values.