Cleans up the cycles to ns conversion in pps-gmtimer.
This replaces the homegrown conversion based on picoseconds per Hz
with one based on the standard clocksource_cyc2ns() function.  In
order to make this work, the clocksource mult/shift fields need to be
set up even when not registering the clocksource for system use.  This
is done via the same clocks_calc_mult_shift() function as is used by
clocksource_register_hz().

TESTED:
With temporary instrumentation, verified that the reported deltas are
consistent with the counter deltas, and more accurate than the old
calculation when they differ (though only by 1ns).