Adds support for setting clock frequencies in pps-gmtimer.
This supports two types of frequency setting:

1) Setting the nominal frequency of external-clock timers, overriding
the device-tree value (if any).  It must be in the plausible range
(currently 1MHz to 24MHz), and the timer must not be the current
clocksource.  This also sets the current frequency.  If this timer
wasn't initially registered as a clocksource (due to the unknown
frequency), an attempt is made to register it now.

2) Setting the current frequency of any timer, but only within 200ppm
of nominal.  This currently does not work for the current-clocksource
timer, due to lack of access to the timekeeping internals.

TESTED:
Verified correct behavior in all cases.