Adds frequency compensation to pps-gmtimer.
This uses the measured frequency rather than the specified frequency
to convert the latency offset into time.  It's based on the ratio
between time and counter deltas between interrupts, but only when the
previous interrupt is within 2^32-1 ns of the current.  When
effective, it avoids an error proportional to the product of the
latency and the frequency error.

TESTED:
With temporary instrumentation, verified that the difference between
this result and the standard result is plausible, including in cases
where the frequency is intentionally specified very incorrectly.  Also
verified that the difference is 0 when the frequency is specified
accurately.