Adds "method" reporting to pps-gmtimer.
This adds a sysfs variable to report the method used to convert the
captured timer value to a timestamp.  The only possible values right
now are "configured_frequency" and "measured_frequency", but this may
get expanded in the future.

TESTED:
Verified proper results, including switching to "configured" on first
PPS event after an interruption.